requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
transformers==4.41.2
sentence-transformers>=2.6.0  # Matches 4.0.2
faiss-cpu>=1.8.0
torch==2.4.1